refactor(direct-control): extract runtime types
Refs: openspec/changes/civ7-support-direct-control-modularization

Moves only runtime API inspection, bounded root inspection, App UI snapshot, Tuner health, playable status, and runtime probe result/input type ownership from packages/civ7-direct-control/src/index.ts into existing runtime atom owner modules while keeping facade type re-exports in index.ts.

Preserves public type contracts and does not move source strings, command execution, session/socket lifecycle, reconnect, setup lifecycle, procedure schemas, telemetry, AI ingestion, hotseat runtime proof, CLI semantic projection, Effect/oRPC procedure-core work, or Task 2.9.4 matrix-row acceptance.

Validation:
- bun run --cwd packages/civ7-direct-control check
- bun run --cwd packages/civ7-direct-control test -- runtime-and-catalog.test.ts public-api.test.ts
- bun run --cwd packages/cli test -- game.control.test.ts
- bun run --cwd packages/civ7-direct-control test
- bun run --cwd packages/civ7-direct-control build
- bun run openspec -- validate civ7-support-direct-control-modularization --strict
- bun run resources:status
- bun run check:cli
- bun run test:cli:play
- bun run --cwd apps/mapgen-studio check
- git diff --check
- ownership/no-cycle rg scans

Local package/type relocation proof only; no runtime/live-game, hotseat product-path, AI-ingestion, semantic CLI, telemetry, or procedure-core proof claimed.
