+++ openspec/changes/civ7-support-direct-control-modularization/workstream/workstream-record.md
@@ -596,6 +596,16 @@ All future agent waves must be framed before delegation:
   semantic projection, Effect/oRPC procedure-core work, or Task 2.9.4
   matrix-row acceptance. This is local package/type relocation proof only, not
   runtime proof, AI/hotseat product-path support, or procedure-core readiness.
+- Direct-control runtime types slice: completed as a narrow public/runtime type
+  ownership relocation. It moves only runtime API inspection, bounded root
+  inspection, App UI snapshot, Tuner health, playable status, and runtime probe
+  result/input types into existing runtime atom owners while keeping facade
+  type re-exports in `index.ts`. This preserves public type contracts and does
+  not move source strings, command execution, session/lifecycle orchestration,
+  procedure schemas, telemetry, AI ingestion, hotseat runtime proof, CLI
+  semantic projection, Effect/oRPC procedure-core work, or Task 2.9.4
+  matrix-row acceptance. This is local package/type relocation proof only, not
+  runtime proof, AI/hotseat product-path support, or procedure-core readiness.
 - Direct-control turn-completion slice: completed as a narrow turn-completion
   source/wrapper relocation. It moves only `getCiv7TurnCompletionStatus`,
   `sendCiv7TurnComplete`, `sendCiv7TurnUnready`, the status command builder,
