+++ packages/civ7-direct-control/src/runtime/probe.ts
@@ -1,0 +1,10 @@
+export type Civ7RuntimeProbe<T> = Readonly<
+  | {
+      ok: true;
+      value: T;
+    }
+  | {
+      ok: false;
+      error: string;
+    }
+>;
